[Feat] UMCT RPC Client/Server Architecture.

+++ Saurons/Radium/src/test/java/com/protobuf/Raccoon.java
@@ -6,7 +6,7 @@ import com.pinecone.hydra.umct.stereotype.Iface;
 //@Iface
 public interface Raccoon {
     @Iface
-    String scratch( String target, long time );
+    String scratch( String target, int time );
 
 //    @Iface( name = "scratchF1" )
 //    default String scratch( String target, long[] times ) {
